Refactor test suite into module-specific files
